Loop video when right slider handle is reached
The Youtube API suggested that we could use the `loop` playerVar to have
this happen automatically, but it didn't seem to behave the way the docs
claimed. So we do it manually by setting the end point to the right
slider handle, and reseeking to the left handle when the video ends.